Allow a custom zk HostProvider
Also create a ZookeeperDefaultModule with all the default zk settings.

cashapp will use this internally to provide a custom HostProvider that
prefers the global zk in the local region.
