+++ misk-zookeeper/src/test/kotlin/misk/clustering/zookeeper/ZkLeaseTestModule.kt
@@ -3,7 +3,6 @@ package misk.clustering.zookeeper
 import com.google.inject.Provides
 import com.google.inject.Singleton
 import misk.clustering.fake.FakeClusterModule
-import misk.clustering.weights.FakeClusterWeight
 import misk.clustering.weights.FakeClusterWeightModule
 import misk.concurrent.ExplicitReleaseDelayQueue
 import misk.config.AppName
@@ -18,7 +17,7 @@ internal class ZkLeaseTestModule : KAbstractModule() {
     bind<String>().annotatedWith<AppName>().toInstance("my-app")
     install(FakeClusterModule())
     install(ZkTestModule(ForZkLease::class))
-    install(ZkLeaseManagerModule())
+    install(ZkLeaseCommonModule())
     install(FakeClusterWeightModule())
   }
